time to practise react and component basics
