+++ src/App.js
@@ -1,38 +1,12 @@
-import ExpenseItem from "./components/ExpenseItem";
+import ExpensesList from "./components/ExpensesList";
 
 const App = () => {
-  const expenses = [
-    {
-      id: 'e1',
-      title: 'Toilet Paper',
-      amount: 94.12,
-      date: new Date(2020, 7, 14),
-    },
-    { id: 'e2', title: 'New TV', amount: 799.49, date: new Date(2021, 2, 12) },
-    {
-      id: 'e3',
-      title: 'Car Insurance',
-      amount: 294.67,
-      date: new Date(2021, 2, 28),
-    },
-    {
-      id: 'e4',
-      title: 'New Desk (Wooden)',
-      amount: 450,
-      date: new Date(2021, 5, 12),
-    },
-  ];
-
   return (
     <div>
       <h2>Let's get started</h2>
-      {
-        expenses.map(({ title, amount, date, id }) => (
-          <ExpenseItem title={title} amount={amount} date={date} key={id} />
-        ))
-      }
+      <ExpensesList />
     </div>
   );
-}
+};
 
 export default App;
